Add services section including WCR (#46)
[DST Guides](https://github.com/DEFRA/dst-guides) is slowly being replaced by [Software development standards](https://github.com/DEFRA/software-development-standards) (SDS) and as such will eventually become defunct.

SDS however won't have a home for the information DST Guides currently contains on specific digital services. Plus it means new members to the **Ruby services team** have to go to multiple places for information on what we cover. Wouldn't if be nice if it was all in one place? 😀

Hence this change, which adds a new section for 'services' and which migrates the existing information on Waste Carriers.