Improve WEX renewal window example
Prior to this change the example used was picked to match the one in our WEX activation and expiration note.

However in light of recent testing it didn't really help to clarify when exactly the 'grace' window started and ended.

Hence this tweak to the example which hopefully makes things clearer.