Update WCR service notes
The current focus of the team is 'tech debt'. We are actively working on migrating functionality out of the old projects ([waste-carriers-frontend](https://github.com/DEFRA/waste-carriers-frontend) and [waste-carriers-service](https://github.com/DEFRA/waste-carriers-service)) and into the new ones.

This means there will soon be changes to how some things are done in Waste Carriers. Because of this we had a quick review of the current notes, and these changes will hopefully bring things up to date.