Add note on FRAE multiple address issue (#78)
https://eaflood.atlassian.net/browse/RIP-359
https://eaflood.atlassian.net/browse/RUBY-866

A quick summary (because it's covered in the note!) but FRAE has a known issue where its possible to create multiple primary organisation addresses when there should only be one.

This has been around awhile (2017!) but came to light again recently when attempting to QA a new EPR export feature for FRAE. To complete verification we needed to test with live data, and having restored the production database it was observed that we were seeing different results to production.

As this was a cause for concern, for example had we broken something else, it led us to investigate and then remind ourselves of this old issue. So it was decided with @andrewhick that we should add this note, so we at least have something to refer back to should we find ourselves again worrying why a search in one environment does not match one in production!